Extract projects list
